feat(ui): add ModuleLayout, TopLoader, extract EY theme, reorganize primitives
- Add nextjs-toploader dependency and TopLoader component
- Create ModuleLayout component (shared layout wrapper with TopLoader + sidebar + main)
- Expand ey-brand.css with full EY theme: palette, shadcn token remapping, animations, utility classes
- Reorganize shadcn components into primitives/ subfolder (avatar, badge, button, card, dialog, input, label, progress, separator, tooltip)
- Keep custom components at root level (sidebar/, module-layout, top-loader)
- Update index.ts exports with grouped sections (primitives vs custom vs utilities)

No downstream import changes needed — all apps import from '@repo/ui' barrel export.
