+++ Frontend/packages/ui/package.json
@@ -38,6 +38,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.469.0",
+    "nextjs-toploader": "^3.9.17",
     "tailwind-merge": "^2.6.0"
   },
   "peerDependencies": {
