+++ Frontend/packages/ui/src/components/primitives/card.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 
 const Card = React.forwardRef<
   HTMLDivElement,
