+++ Frontend/packages/ui/src/components/primitives/input.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 
 const Input = React.forwardRef<
   HTMLInputElement,
