+++ Frontend/packages/ui/src/components/primitives/label.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../lib/utils";
+import { cn } from "../../lib/utils";
 
 const Label = React.forwardRef<
   HTMLLabelElement,
