+++ Frontend/packages/ui/src/components/top-loader.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import NextTopLoader from "nextjs-toploader";
+
+export function TopLoader() {
+  return <NextTopLoader color="#2d2d2d" height={3} showSpinner={false} />;
+}
